Improve doc and run scripts

+++ onboarding/run.sh
@@ -2,5 +2,8 @@
 rm pulumi.log || true
 ./run_tests.sh "$@"
 
+#Exit 8 custom output for run_tests.sh == Invalid parameter
+[[ "$?" == "8" ]] && exit 1
+
 # .:: Destroy infraestructure ::.
 aws-vault exec sandbox-account-admin -- pulumi destroy --yes -C . -s dev
